Fix Canva sponsor div alignment
Fixes an issue with rendering of certain sponsor logos. Because the parent column is a col-md-12, having two child divs be col-md-5 instead of col-md-6 left them slightly off centre. This PR corrects this issue. I've also taken the liberty of adding a link to the Canva image (because I know where it should be directed), I'd suggest you do the same for the others (as I can't speak to where they should be linked, though I'd suggest a careers page, or failing that just the sponsor home page).